+++ .github/workflows/linters.yml
@@ -48,7 +48,7 @@ jobs:
         uses: liskin/gh-problem-matcher-wrap@d8afa2cfb66dd3f982b1950429e652bc14d0d7d2
         with:
           linters: isort
-          run: isort --check --diff re_position products
+          run: isort --check --diff re_position product
 
   black:
     runs-on: ubuntu-latest
